requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
daikinapi==1.0.5
idna==2.9                 # via requests
prometheus-client==0.7.1
python-dotenv==0.12.0
requests==2.23.0          # via daikinapi
urllib3==1.25.8           # via daikinapi, requests
